requirements.txt
APScheduler==3.6.3
astor==0.8.1
attrs==20.3.0
bandit==1.7.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
darglint==1.8.0
docutils==0.17.1
eradicate==2.0.0
flake8==3.9.1
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==20.11.1
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-eradicate==1.0.0
flake8-isort==4.0.0
flake8-polyfill==1.0.2
flake8-quotes==3.2.0
flake8-rst-docstrings==0.0.14
flake8-string-format==0.3.0
gitdb==4.0.7
GitPython==3.1.14
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
isort==5.8.0
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
pbr==5.5.1
pep8-naming==0.11.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydocstyle==6.0.0
pyflakes==2.3.1
Pygments==2.8.1
pyparsing==2.4.7
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
restructuredtext-lint==1.3.2
rsa==4.7.2
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
stevedore==3.3.0
testfixtures==6.17.1
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.4
wemake-python-styleguide==0.15.2
